requirements.txt
Flask
requests
google-cloud-dialogflow
google-cloud-translate
python-dotenv
gunicorn
google-generativeai>=0.6.0
pandas
qdrant-client
qdrant-client
python-docx
pymupdf
requests
tqdm